.claude/worktrees/sad-haslett/requirements.txt
# Hackathon Completion Engine - Dependencies
# Python 3.11+ required

# Core Dependencies
pydantic==2.5.0          # Data validation
jinja2==3.1.2            # Template rendering
python-dotenv==1.0.0     # Environment variables
requests==2.31.0         # HTTP requests
pyyaml==6.0.1            # YAML parsing
click==8.1.7             # CLI framework
rich==13.7.0             # Terminal formatting
flask==2.3.3             # Web framework for webhook server
discord.py==2.3.2        # Discord API wrapper

# Development Dependencies
pytest==7.4.3            # Testing
mypy==1.7.1              # Type checking
pylint==3.0.3            # Linting
black==23.12.0           # Code formatting
